+++ proofs/proof.mli
@@ -244,13 +244,6 @@ module V82 : sig
   (* returns the existential variable used to start the proof *)
   val top_evars : t -> Evar.t list
 
-  (* Turns the unresolved evars into goals.
-     Raises [UnfinishedProof] if there are still unsolved goals. *)
-  val grab_evars : t -> t
-
-  (* Implements the Existential command *)
-  val instantiate_evar :
-    Environ.env -> int -> Constrexpr.constr_expr -> t -> t
 end
 
 (* returns the set of all goals in the proof *)